% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_dimension.R
\name{box_dimension}
\alias{box_dimension}
\title{Box Dimension}
\usage{
box_dimension(
  cloud,
  x = 1,
  y = 2,
  z = 3,
  lowercutoff = 0.01,
  rm_int_box = FALSE
)
}
\arguments{
\item{cloud}{A point cloud object}

\item{x}{The column name or position of the x coordinates. Defaults to the first column.}

\item{y}{The column name or position of the y coordinates. Defaults to the second column.}

\item{z}{The column name or position of the x coordinates. Defaults to the third column.}

\item{lowercutoff}{The smallest box size determined by the point spacing of the cloud in meters. Defaults to 1 cm.}

\item{rm_int_box}{Remove the initial box as TRUE or FALSE. Defaults to FALSE.}
}
\value{
Returns a list
}
\description{
R port of Dominik Seidel's fractal analysis "box-dimension" metric.
}
\examples{
\dontrun{
## Calculate Box Dimension
file <- system.file("extdata/cloud.txt", package = "rTwig")
cloud <- read.table(file)
output <- box_dimension(cloud)
output

## Plot Results
data <- output[[1]]
plot(data$log.box.size,
  data$log.voxels,
  pch = 19,
  xlab = "Log(Inverse Box Size)",
  ylab = "log(Box Count)"
)
abline(lm(data$log.voxels ~ data$log.box.size))
}
}
\references{
{
\insertRef{box_dimension1}{rTwig}

\insertRef{box_dimension2}{rTwig}
}
}
