% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_twigs.R
\name{download_twigs}
\alias{download_twigs}
\title{Download Twigs}
\usage{
download_twigs(database = "twigs")
}
\arguments{
\item{database}{Twig database to return. Can be "twigs", "twigs_index",
"raw", for the raw twig measurements, or "all" for all databases.}
}
\value{
Returns a data frame
}
\description{
Download twig measurements from an external database, including
additional, verified measurements from contributors.
}
\examples{

download_twigs()

}
