% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_blender.R
\name{export_blender}
\alias{export_blender}
\title{Export Blender}
\usage{
export_blender(cylinder, filename)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{filename}{File name and path for exporting. The .txt extension is automatically added if not present.}
}
\value{
A .txt file
}
\description{
Exports a QSM in the QSM Blender Addons format
}
\examples{

file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- run_rtwig(file, twig_radius = 4.23, metrics = FALSE, smooth = TRUE)

filename <- tempfile(pattern = "qsm_blender")
export_blender(qsm, filename)

}
