% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_length.R
\name{growth_length}
\alias{growth_length}
\title{Growth Length}
\usage{
growth_length(cylinder)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}
}
\value{
Returns a data frame
}
\description{
Calculates the growth length of each cylinder
}
\details{
Growth length is the length of a parent cylinder, plus the lengths
of all of its child cylinders. It is a cumulative metric that was created by
Dr. Jan Hackenberg for use in the SimpleForest QSM software.
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
str(cylinder)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
str(cylinder)
}
}
\references{
{
\insertRef{growth_parameter1}{rTwig}
}
}
