% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qsm.R
\name{import_adqsm}
\alias{import_adqsm}
\title{Import AdQSM}
\usage{
import_adqsm(...)
}
\arguments{
\item{...}{function inputs}
}
\value{
Returns a data frame
}
\description{
\code{import_adqsm()} is deprecated and will be removed in a future
version. Use \code{import_qsm()} instead.
}
