% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_leaves.R
\name{import_leaves}
\alias{import_leaves}
\title{Import Leaves}
\usage{
import_leaves(filename, simplify = FALSE)
}
\arguments{
\item{filename}{a QSM-FaNNI .obj file}

\item{simplify}{simplify the mesh by removing duplicate vertices. This can
reduce memory usage at the cost of import speed. Defaults to FALSE.}
}
\value{
Returns an RGL mesh3d object
}
\description{
Imports leaves created by the QSM foliage and needles naive
insertion (QSM-FaNNI) algorithm. The leaves are stored as a triangular mesh,
compatible with both rTwig and RGL functions.
}
\examples{
\dontrun{

## Import FaNNI leaves from the standard obj format
file <- "path_to_leaves.obj"
qsm <- import_treeqsm(file)

## Import FaNNI leaves from the extended obj format
# ADD THIS FUNCTIONALITY!!!
file <- "path_to_leaves.obj"
qsm <- import_treeqsm(file)
}

}
\references{
\insertRef{FaNNI}{rTwig}
}
