% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qsm.R
\name{import_treegraph}
\alias{import_treegraph}
\title{Import Treegraph}
\usage{
import_treegraph(...)
}
\arguments{
\item{...}{function inputs}
}
\value{
Returns a list
}
\description{
\code{import_treegraph()} is deprecated and will be removed in a future
version. Use \code{import_qsm()} instead.
}
