% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qsm.R
\name{import_treeqsm}
\alias{import_treeqsm}
\title{Import TreeQSM}
\usage{
import_treeqsm(...)
}
\arguments{
\item{...}{function inputs}
}
\value{
Returns a list
}
\description{
\code{import_treeqsm()} is deprecated and will be removed in a future
version. Use \code{import_qsm()} instead.
}
