% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qsm_cloud.R
\name{plot_qsm_cloud}
\alias{plot_qsm_cloud}
\title{Plot QSM and Point Cloud}
\usage{
plot_qsm_cloud(
  cylinder,
  cloud,
  radius = "modified",
  cyl_color = "BranchOrder",
  pt_color = "black",
  pt_size = 0.1
)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{cloud}{Point cloud data frame where the first three columns are the x, y, and z coordinates in the same coordinate system as the QSM.}

\item{radius}{Radius type, as either "modified" or "unmodified". Defaults to modified.}

\item{cyl_color}{Color QSM by either "BranchOrder" or "GrowthLength". Defaults to BranchOrder.}

\item{pt_color}{Color of the point cloud. Defaults to black.}

\item{pt_size}{Size of the points. Defaults to 0.1.}
}
\value{
A rgl QSM plot with overlayed point cloud
}
\description{
Plots QSM cylinders and point cloud using the rgl library
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)

file2 <- system.file("extdata/cloud.txt", package = "rTwig")
cloud <- read.table(file2, header = FALSE, sep = ",")

plot_qsm_cloud(cylinder, cloud)
)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)

file2 <- system.file("extdata/cloud.txt", package = "rTwig")
cloud <- read.table(file2, header = FALSE, sep = ",")

plot_qsm_cloud(
  cylinder = cylinder,
  cloud = cloud,
  radius = "unmodified",
  cyl_color = "GrowthLength",
  pt_color = "blue",
  pt_size = 0.2
)
}
}
