% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twigs.R
\docType{data}
\name{twigs}
\alias{twigs}
\title{Tree Twig Radii Data}
\format{
\subsection{\code{twigs}}{

A data frame with 34 rows and 3 columns:
\describe{
\item{scientific.name}{The tree's genus and species}
\item{SpCode}{The two letter FIA Species Code}
\item{radius.m}{The average twig radius in meters}
}
}
}
\usage{
twigs
}
\description{
Average twig radii for common North American tree species
}
\keyword{datasets}
