% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twigs.R
\docType{data}
\name{twigs_index}
\alias{twigs_index}
\title{Twig Index Database}
\format{
\subsection{\code{twigs_index}}{

A data frame containing twig radii measurements
\describe{
\item{size_index}{The twig size index}
\item{radius_mm}{The average twig radius in millimeters}
\item{n}{The twig measurement sample size}
\item{min}{The minimum twig radii from the samples}
\item{max}{The maximum twig radii from the samples}
\item{std}{The standard deviation of twig radii}
\item{cv}{The coefficient of variation of twig radii}
}
}
}
\usage{
twigs_index
}
\description{
Database of twig radii by size index
}
\keyword{datasets}
