% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_radii.R
\name{correct_radii}
\alias{correct_radii}
\title{Correct Radii}
\usage{
correct_radii(df, twigRad)
}
\arguments{
\item{df}{QSM cylinder data frame}

\item{twigRad}{Twig radius in meters.}
}
\value{
Returns a data.frame
}
\description{
Corrects cylinder radii.
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
df <- update_cylinders(df)
df <- growth_length(df)
df <- correct_radii(df, twigRad = 0.003)
str(df)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
df <- read.csv(file)
df <- update_cylinders(df)
df <- growth_length(df)
df <- correct_radii(df, twigRad = 0.003)
str(df)
}
}
