% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_radii.R
\name{correct_radii}
\alias{correct_radii}
\title{Correct Radii}
\usage{
correct_radii(cylinder, twigRad)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{twigRad}{Twig radius in millimeters}
}
\value{
Returns a data frame
}
\description{
Corrects cylinder radii
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
cylinder <- correct_radii(cylinder, twigRad = 1.5)
str(cylinder)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
cylinder <- correct_radii(cylinder, twigRad = 1.5)
str(cylinder)
}
}
