% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mat.R
\name{export_mat}
\alias{export_mat}
\title{Export MAT}
\usage{
export_mat(
  cylinder,
  filename,
  metrics = TRUE,
  rundata = NULL,
  pmdistance = NULL,
  triangulation = NULL
)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{filename}{Desired name of file}

\item{metrics}{Calculate treedata and branch structs? Defaults to TRUE.}

\item{rundata}{Optional rundata list created by \code{import_treeqsm()}}

\item{pmdistance}{Optional pmdistance list created by \code{import_treeqsm()}}

\item{triangulation}{Optional triangulation list created by \code{import_treeqsm()}}
}
\value{
Returns a .mat file
}
\description{
Exports the cylinder data to the format used by TreeQSM
}
\examples{

## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_treeqsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)

filename <- tempfile(pattern = "TreeQSM_QSM", fileext = ".mat")
export_mat(cylinder, filename)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)

filename <- tempfile(pattern = "SimpleForest_QSM", fileext = ".mat")
export_mat(cylinder, filename)

}
