% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mat.R
\name{export_mat}
\alias{export_mat}
\title{Export MAT}
\usage{
export_mat(df, filename, method = "TreeQSM")
}
\arguments{
\item{df}{QSM cylinder data frame}

\item{filename}{Desired name of file}

\item{method}{QSM type, as either "TreeQSM" or "SimpleForest". Defaults to TreeQSM.}
}
\value{
Returns a .mat file
}
\description{
Exports the cylinder data to be visualized with TreeQSM's plot_cylinder_model() function.
}
\examples{
\dontrun{
## TreeQSM Processing Chain
df <- import_qsm("foo.mat")
df <- update_cylinders(df)
df <- growth_length(df)
df <- correct_radii(df, twigRad = 0.003)
export_mat(df, "foo_corrected.mat")

## SimpleForest Processing Chain
df <- read.csv("foo.csv")
df <- update_cylinders(df, method = "SimpleForest")
df <- growth_length(df, method = "SimpleForest")
df <- correct_radii(df, twigRad = 0.003, method = "SimpleForest")
export_mat(df, "foo_corrected.mat", method = "SimpleForest")
}
}
