% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mat.R
\name{export_mat}
\alias{export_mat}
\title{Export MAT}
\usage{
export_mat(cylinder, filename)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{filename}{Desired name of file}
}
\value{
Returns a .mat file
}
\description{
Exports the cylinder data to be visualized with TreeQSM's plot_cylinder_model() function
}
\examples{
\dontrun{
## TreeQSM Processing Chain
qsm <- import_qsm("foo.mat")
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
cylinder <- correct_radii(cylinder, twigRad = 1.5)
export_mat(cylinder, "foo_corrected.mat")

## SimpleForest Processing Chain
cylinder <- read.csv("foo.csv")
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
cylinder <- correct_radii(cylinder, twigRad = 1.5)
export_mat(cylinder, "foo_corrected.mat")
}
}
