% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_length.R
\name{growth_length}
\alias{growth_length}
\title{Growth Length}
\usage{
growth_length(df)
}
\arguments{
\item{df}{QSM cylinder data.frame}
}
\value{
Returns a data.frame
}
\description{
Calculates the growth length of each cylinder.
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
df <- update_cylinders(df)
df <- growth_length(df)
str(df)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
df <- read.csv(file)
df <- update_cylinders(df)
df <- growth_length(df)
str(df)
}
}
