% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_length.R
\name{growth_length}
\alias{growth_length}
\title{Growth Length}
\usage{
growth_length(cylinder)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}
}
\value{
Returns a data frame
}
\description{
Calculates the growth length of each cylinder
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
str(cylinder)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
str(cylinder)
}
}
