% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_adqsm.R
\name{import_adqsm}
\alias{import_adqsm}
\title{Import AdQSM}
\usage{
import_adqsm(filename, method = "adqsm")
}
\arguments{
\item{filename}{an AdQSM or AdTree .obj file}

\item{method}{specify the method as \code{adqsm} or \code{adtree}. Defaults to \code{adqsm}.}
}
\value{
Returns a data frame
}
\description{
Imports a QSM created by AdQSM or AdTree
}
\examples{
\dontrun{

# Import an AdQSM or AdTree
qsm <- import_adqsm("path/to/obj/file")
}

}
\references{
\insertRef{fan2020adqsm}{rTwig}
\insertRef{du2019adtree}{rTwig}
}
