% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_adqsm.R
\name{import_adqsm}
\alias{import_adqsm}
\title{Import AdQSM}
\usage{
import_adqsm(filename)
}
\arguments{
\item{filename}{an AdQSM .obj file}
}
\value{
Returns a list
}
\description{
Imports a QSM created by AdQSM
}
\examples{
\dontrun{

# Import an AdQSM
qsm <- import_adqsm("path/to/obj/file")
}

}
\references{
\insertRef{fan2020adqsm}{rTwig}
}
