% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qsm.R
\name{import_qsm}
\alias{import_qsm}
\title{Import QSM}
\usage{
import_qsm(filename)
}
\arguments{
\item{filename}{a QSM file. Accepts \code{.mat} (TreeQSM, SmartQSM),
\code{.json} (Treegraph), or \code{.obj} (AdQSM, AdTree)}
}
\value{
Returns a list
}
\description{
Imports a QSM created by TreeQSM, Treegraph, AdQSM, AdTree,
or SmartQSM
}
\examples{

# Read a TreeQSM MATLAB file
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
summary(qsm)

}
\references{
\insertRef{TreeQSM}{rTwig}
\insertRef{treegraph2}{rTwig}
\insertRef{treegraph1}{rTwig}
\insertRef{fan2020adqsm}{rTwig}
\insertRef{du2019adtree}{rTwig}
\insertRef{SmartQSM}{rTwig}
}
