% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qsm.R
\name{import_qsm}
\alias{import_qsm}
\title{Import QSM}
\usage{
import_qsm(file)
}
\arguments{
\item{file}{a .mat file}
}
\value{
Returns a data.frame
}
\description{
Imports a QSM stored as a MATLAB file created by TreeQSM.
}
\examples{
## Read a TreeQSM MATLAB file
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
str(df)
}
