% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_treeqsm.R
\name{import_qsm}
\alias{import_qsm}
\title{Import TreeQSM}
\usage{
import_qsm(...)
}
\arguments{
\item{...}{function inputs}
}
\value{
Returns a list
}
\description{
\code{import_qsm()} is deprecated and will be removed in a future version. Use \code{import_treeqsm()} instead.
}
