% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_treegraph.R
\name{import_treegraph}
\alias{import_treegraph}
\title{Import Treegraph}
\usage{
import_treegraph(file)
}
\arguments{
\item{file}{a treegraph .json file}
}
\value{
Returns a list
}
\description{
Imports a QSM created by treegraph
}
\examples{
\dontrun{

# Import a treegraph QSM
qsm <- import_treegraph("path/to/json/file")
}

}
\references{
\insertRef{treegraph2}{rTwig}

\insertRef{treegraph1}{rTwig}
}
