% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_treeqsm.R
\name{import_treeqsm}
\alias{import_treeqsm}
\title{Import TreeQSM}
\usage{
import_treeqsm(filename)
}
\arguments{
\item{filename}{a TreeQSM .mat MATLAB file}
}
\value{
Returns a list
}
\description{
Imports a QSM created by TreeQSM
}
\examples{

## Read a TreeQSM MATLAB file in the 2.3.x - 2.4.x format
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_treeqsm(file)
summary(qsm)

## Read a TreeQSM MATLAB file in the 2.0 format
file <- system.file("extdata/QSM_2.mat", package = "rTwig")
qsm <- import_treeqsm(file)
names(qsm)

}
\references{
\insertRef{TreeQSM}{rTwig}
}
