% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qsm.R
\name{plot_qsm}
\alias{plot_qsm}
\title{Plot QSM}
\usage{
plot_qsm(cylinder, radius = "modified", color = "BranchOrder")
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{radius}{Radius type, as either "modified" or "unmodified". Defaults to modified.}

\item{color}{Color QSM by either "BranchOrder" or "GrowthLength". Defaults to BranchOrder.}
}
\value{
A rgl QSM plot
}
\description{
Plots QSM cylinders using the rgl library
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
cylinder <- growth_length(cylinder)
cylinder <- correct_radii(cylinder, twigRad = 1.5)
plot_qsm(cylinder)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
cylinder <- correct_radii(cylinder, twigRad = 1.5)
plot_qsm(cylinder, radius = "unmodified", color = "GrowthLength")
}
}
