% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qsm.R
\name{plot_qsm}
\alias{plot_qsm}
\title{Plot QSM}
\usage{
plot_qsm(
  cylinder = NULL,
  radius = NULL,
  color = NULL,
  palette = NULL,
  facets = 6,
  skeleton = FALSE,
  skeleton_lwd = NULL,
  cloud = NULL,
  pt_color = NULL,
  pt_size = NULL,
  axes = TRUE,
  axes_color = NULL,
  hover = FALSE,
  bg_color = NULL,
  lit = TRUE,
  pan = TRUE
)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{radius}{Column name of radii as a quoted string. Defaults to the modified radii.}

\item{color}{Optional cylinder color parameter. Colors must be a single hex color, a vector of hex colors, or a quoted column name. It can also be set to "random" to generate a random solid color. Vectors must have the same length as the cylinder data frame.}

\item{palette}{Optional color palette for numerical data. Palettes include: viridis, inferno, plasma, magma, cividis, and rainbow.}

\item{facets}{The number of facets in the polygon cross section. Defaults to 6, but can be increased to improve visual smoothness at the cost of performance and memory.}

\item{skeleton}{Plot the QSM skeleton instead of cylinders. Defaults to FALSE.}

\item{skeleton_lwd}{Skeleton line width. Defaults to 1.}

\item{cloud}{Point cloud data frame where the first three columns are the x, y, and z coordinates in the same coordinate system as the QSM. Defaults to NULL.}

\item{pt_color}{Color of the point cloud. Defaults to black. Can be set to "random".}

\item{pt_size}{Size of the points. Defaults to 0.1.}

\item{axes}{Show plot axes. Defaults to TRUE.}

\item{axes_color}{Set the axes color. Defaults to black.}

\item{hover}{Show cylinder and branch id on mouse hover. Defaults to FALSE.}

\item{bg_color}{Set the background color of the plot. Defaults to white.}

\item{lit}{Enable light source in plot. Defaults to TRUE. Can be set to FALSE.}

\item{pan}{Use right mouse button to pan plot. Defaults to TRUE, but is disabled when hover is enabled.}
}
\value{
A rgl plot
}
\description{
Efficiently plot QSMs and point clouds. Uses the Rcpp and RGL libraries as backends.
}
\examples{

## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
plot_qsm(cylinder)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
plot_qsm(cylinder)

## aRchi Processing Chain
file <- system.file("extdata/QSM2.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
plot_qsm(cylinder)

}
