% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qsm.R
\name{plot_qsm}
\alias{plot_qsm}
\title{Plot QSM}
\usage{
plot_qsm(df, radius = "modified", color = "BranchOrder", method = "TreeQSM")
}
\arguments{
\item{df}{QSM cylinder data frame}

\item{radius}{Radius type, as either "modified" or "unmodified". Defaults to modified.}

\item{color}{Color QSM by either "BranchOrder" or "GrowthLength". Defaults to BranchOrder.}

\item{method}{QSM type, as either "TreeQSM" or "SimpleForest". Defaults to TreeQSM.}
}
\value{
A rgl QSM plot
}
\description{
Plot QSM
}
\examples{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
df <- update_cylinders(df)
df <- growth_length(df)
plot_qsm(df)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
df <- read.csv(file)
df <- update_cylinders(df, method = "SimpleForest")
df <- growth_length(df, method = "SimpleForest")
plot_qsm(df, radius = "unmodified", color = "GrowthLength", method = "SimpleForest")
}
