% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsm_summary.R
\name{qsm_summary}
\alias{qsm_summary}
\title{QSM Summary}
\usage{
qsm_summary(df, method = "TreeQSM")
}
\arguments{
\item{df}{QSM cylinder data frame}

\item{method}{QSM type, as either "TreeQSM" or "SimpleForest". Defaults to TreeQSM.}
}
\value{
Returns a list
}
\description{
Generates summaries of QSM features (e.g. volume, surface area, dbh, etc.) by totals and branch order.
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
summary <- qsm_summary(df)
str(summary)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
df <- read.csv(file)
summary <- qsm_summary(df, method = "SimpleForest")
str(summary)
}
}
