% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_qsm().R
\name{reconstruct_qsm}
\alias{reconstruct_qsm}
\title{Reconstruct QSM}
\usage{
reconstruct_qsm(
  cylinder,
  id,
  parent,
  radius,
  start_x,
  start_y,
  start_z,
  end_x = NULL,
  end_y = NULL,
  end_z = NULL,
  axis_x = NULL,
  axis_y = NULL,
  axis_z = NULL,
  length = NULL,
  branch = NULL,
  branch_order = NULL,
  metrics = FALSE
)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}

\item{id}{cylinder ids}

\item{parent}{cylinder parent ids}

\item{radius}{cylinder radius}

\item{start_x}{cylinder start x position}

\item{start_y}{cylinder start y position}

\item{start_z}{cylinder start z position}

\item{end_x}{cylinder end x position}

\item{end_y}{cylinder end y position}

\item{end_z}{cylinder end z position}

\item{axis_x}{cylinder end x position}

\item{axis_y}{cylinder end y position}

\item{axis_z}{cylinder end z position}

\item{length}{cylinder length}

\item{branch}{optional cylinder branch id}

\item{branch_order}{optional cylinder branch order}

\item{metrics}{Calculate tree metrics. Defaults to FALSE.}
}
\value{
data frame
}
\description{
Reconstruct a QSM and all of its variables from the minimum amount of
required data. The required variables are the cylinder id, parent id,
3d information and the radius. The cylinder can be defined in 3d space with
either a combination of the start and end points, or the start, axis, and
length. Branch and branch order information are optional. If not provided,
the branches and orders are defined recursively using the growth length.
}
\examples{
file <- system.file("extdata/QSM.csv", package = "rTwig")
qsm <- reconstruct_qsm(
  cylinder = read.csv(file),
  id = "ID", parent = "parentID", radius = "radius",
  start_x = "startX", start_y = "startY", start_z = "startZ",
  end_x = "endX", end_y = "endY", end_z = "endZ"
)
plot_qsm(qsm)

}
