% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTwig-package.R
\docType{package}
\name{rTwig-package}
\alias{rTwig}
\alias{rTwig-package}
\title{rTwig: Realistic Quantitative Structure Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Real Twig is a method to correct branch overestimation in quantitative structure models. Overestimated cylinders are correctly tapered using measured twig diameters of corresponding tree species. Supported quantitative structure modeling software includes 'TreeQSM', 'SimpleForest' and 'Treegraph'. Also included is a novel database of twig diameters and tools for fractal analysis of point clouds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aidanmorales.github.io/rTwig/}
  \item \url{https://github.com/aidanmorales/rTwig}
  \item Report bugs at \url{https://github.com/aidanmorales/rTwig/issues}
}

}
\author{
\strong{Maintainer}: Aidan Morales \email{moral169@msu.edu} (\href{https://orcid.org/0009-0000-4513-4908}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item David W. MacFarlane \email{macfar24@msu.edu} [copyright holder]
}

}
\keyword{internal}
