% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_qsm.R
\name{smooth_qsm}
\alias{smooth_qsm}
\title{Smooth QSM}
\usage{
smooth_qsm(df, method = "TreeQSM")
}
\arguments{
\item{df}{QSM cylinder data frame}

\item{method}{QSM type, as either "TreeQSM" or "SimpleForest". Only TreeQSM is currently implemented and is the default.}
}
\value{
Returns a data.frame
}
\description{
Visual smoothing of a QSM by ensuring the midpoints of all branch cylinders are connected.
}
\examples{
\dontrun{
## TreeQSM Smoothing
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
df <- update_cylinders(df)
df <- growth_length(df)
df <- correct_radii(df, twigRad = 0.003)

## Before Smoothing
plot_qsm(df)

## After Smoothing
df <- smooth_qsm(df)
plot_qsm(df)
}
}
