% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twigs.R
\docType{data}
\name{twigs}
\alias{twigs}
\title{Twig Database}
\format{
\subsection{\code{twigs}}{

A data frame containing twig radii measurements
\describe{
\item{scientific.name}{The tree's genus and species}
\item{radius.mm}{The average twig radius in millimeters}
\item{n}{The twig measurement sample size}
\item{min}{The minimum twig radii from the samples}
\item{max}{The maximum twig radii from the samples}
\item{std}{The standard deviation of twig radii}
\item{cv}{The coefficient of variation of twig radii}
}
}
}
\usage{
twigs
}
\description{
Database of twig radii for common North American tree species
}
\keyword{datasets}
