% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cylinders.R
\name{update_cylinders}
\alias{update_cylinders}
\title{Update cylinder}
\usage{
update_cylinders(cylinder)
}
\arguments{
\item{cylinder}{QSM cylinder data frame}
}
\value{
Returns a data frame
}
\description{
Updates the QSM cylinder data in preparation for growth length and path analysis calculations
}
\examples{
\dontrun{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
str(cylinder)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
str(cylinder)
}
}
