% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cylinders.R
\name{update_cylinders}
\alias{update_cylinders}
\title{Update Cylinders}
\usage{
update_cylinders(df, method = "TreeQSM")
}
\arguments{
\item{df}{QSM cylinder data.frame}

\item{method}{QSM type, as either "TreeQSM" or "SimpleForest". Defaults to TreeQSM.}
}
\value{
Returns a data.frame
}
\description{
Updates the QSM cylinder data in preparation for growth length and path analysis calculations.
}
\examples{
\donttest{
## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
df <- import_qsm(file)
df <- update_cylinders(df)
str(df)

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
df <- read.csv(file)
df <- update_cylinders(df, method = "SimpleForest")
str(df)
}
}
